frontend/requirements.txt
aerich==0.6.1
aiosqlite==0.17.0
anyio==3.4.0
asgiref==3.4.1
asyncpg==0.25.0
bcrypt==3.2.0
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.10
click==8.0.3
colorama==0.4.4
cryptography==36.0.1
ddlparse==1.10.0
dictdiffer==0.9.0
dnspython==2.1.0
ecdsa==0.17.0
email-validator==1.1.3
fastapi==0.70.1
h11==0.12.0
httpcore==0.14.4
httpx==0.21.3
idna==3.3
iso8601==0.1.16
Jinja2==3.0.3
MarkupSafe==2.0.1
passlib==1.7.4
pyasn1==0.4.8
pycparser==2.21
pydantic==1.9.0
pyparsing==3.0.6
pypika-tortoise==0.1.2
python-dotenv==0.19.2
python-http-client==3.3.4
python-jose==3.3.0
python-multipart==0.0.5
pytz==2021.3
rfc3986==1.5.0
rsa==4.8
sendgrid==6.9.3
six==1.16.0
sniffio==1.2.0
starkbank-ecdsa==2.0.3
starlette==0.16.0
tomlkit==0.8.0
tortoise-orm==0.17.8
typing_extensions==4.0.1
uvicorn==0.16.0
